+++ release_notes.md
@@ -3,6 +3,7 @@
 - Ephys pipeline: spike sorting
     - dsp.voltage: destripe detects bad channels and interpolate them
     - synchronisation 3B: gives a 10% leeway to throw exception
+    - spikes sorting task outputs the RMS plot
 
 ## Release Notes 2.4
 ### Release Notes 2.4.0